Refactor test

+++ src/ILock.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.13;
 
 interface ILock {
-    event Withdrawal(uint amount, uint when);
+    event Withdrawal(uint256 amount, uint256 when);
 
     function withdraw() external;
 }
